# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API к ИС Не-Ной",
        "title": "Ne-Noy-Api",
        "contact": {},
        "version": "1.0"
    },
    "host": "https://simply-funny-spearfish.cloudpub.ru",
    "basePath": "/api",
    "paths": {
...
        "/v1/events/team": {
            "post": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "description": "Создает запись мероприятия типа \"команды\" с настройками количества команд и вместимости.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Создать командное мероприятие",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные командного мероприятия",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/team_dto.CreateTeamEventDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/team_dto.TeamEventDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/events/team/{id}": {
            "get": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "description": "Возвращает командное мероприятие",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Получить командное мероприятие (по маршруту /events/team)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID командного мероприятия",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/team_dto.TeamEventDto"
                        }
                    },
                    "400": {
                        "description": "Некорректный UUID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Мероприятие не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "description": "Создает новую команду в мероприятии типа \"команды\"; капитаном назначается текущий авторизованный пользователь.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Создать команду в командном мероприятии",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID командного мероприятия",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные команды",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/team_dto.CreateTeamRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/team_dto.TeamDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Мероприятие или пользователь не найдены",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "description": "Удаляет запись мероприятия типа \"команды\" вместе с командами и участниками команд.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Удалить командное мероприятие",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID командного мероприятия",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Некорректный UUID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Командное мероприятие не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "description": "Частично обновляет запись мероприятия типа \"команды\".",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Обновить командное мероприятие",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID командного мероприятия",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Поля для обновления командного мероприятия",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/team_dto.UpdateTeamEventDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/team_dto.TeamEventDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Командное мероприятие не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/events/team/{id}/joinTo/{teamId}": {
            "post": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "description": "Добавляет текущего авторизованного пользователя в выбранную команду командного мероприятия.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Вступить в команду",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID командного мероприятия",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID команды",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Некорректный UUID или команда заполнена",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Команда или пользователь не найдены",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "team_dto.CreateTeamEventDto": {
            "type": "object",
            "properties": {
                "additional_address": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "attachments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "cover": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "organizers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "starts_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "teams_cap_max": {
                    "type": "integer"
                },
                "teams_cap_min": {
                    "type": "integer"
                },
                "teams_constraint": {
                    "type": "integer"
                },
                "vk_post_id": {
                    "type": "integer"
                }
            }
        },
        "team_dto.CreateTeamRequestDto": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "team_dto.TeamEventDto": {
            "type": "object",
            "properties": {
                "additional_address": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "organizers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserMiniDto"
                    }
                },
                "starts_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "teams_cap_max": {
                    "type": "integer"
                },
                "teams_cap_min": {
                    "type": "integer"
                },
                "teams_constraint": {
                    "type": "integer"
                },
                "vk_post_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "VkAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}